2025-02-02/tts/requirements.txt
openai>=1.6.1
python-dotenv==1.0.0
loguru==0.7.2 